retentionrate/requirements.txt
numpy==1.16.6
pandas==0.24.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
six==1.14.0
